requirements.txt
discord.py[voice]
edge-tts
flask